Small Fixes + Add of third Confirm Page
Because:
- preparing for pull request
- better overview over the data in study creation

Fixed the back button bug and changed a few strings to fit better.
Added a third page for comfirmation to have a smoother experience
which is less cluttered.
