+++ app/src/main/java/com/example/vpmanager/createStudyFragment_StepOne_Contact.java
@@ -12,7 +12,7 @@ public class createStudyFragment_StepOne_Contact extends Fragment {
 
 
     public createStudyFragment_StepOne_Contact() {
-        createStudyBase.currentFragment = 2;
+        createStudyActivity.currentFragment = 2;
     }
 
     @Override
