+++ app/src/main/java/com/example/vpmanager/createStudyFragment_StepThree.java
@@ -15,7 +15,7 @@ public class createStudyFragment_StepThree extends Fragment {
     Spinner executionType;
 
     public createStudyFragment_StepThree() {
-        createStudyBase.currentFragment = 4;
+        createStudyActivity.currentFragment = 4;
     }
 
     @Override
