+++ app/src/main/res/layout/fragment_create_study_final_step_three.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".createStudyFragment_finalStep_two"
+    tools:context=".createStudyFragment_finalStep_three"
     android:background="?android:windowBackground">
 
 
@@ -11,13 +11,13 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        android:weightSum="10">
+        android:weightSum="7">
 
         <TextView
             android:id="@+id/pageTitle"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="@string/fragment_create_study_pageFinalTitle"
+            android:text="Terminbestätigung"
             android:textAlignment="center"
             android:textSize="35sp" />
 
@@ -29,43 +29,28 @@
             android:layout_marginRight="20dp"
             android:background="@color/black" />
 
+
         <com.google.android.material.card.MaterialCardView
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_weight="6"
+            android:layout_weight="6.5"
 
             android:layout_marginRight="30dp"
             android:layout_marginLeft="30dp"
             android:layout_marginTop="20dp">
 
-            <TextView
+            <ListView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:id="@+id/confirmDesc"
+                android:text="Dates"
+                android:id="@+id/confirmDates"
                 android:textSize="20sp"
-                android:gravity="top"
-                android:layout_marginStart="10dp"
-                android:layout_marginTop="5dp">
-            </TextView>
-        </com.google.android.material.card.MaterialCardView>
+                android:layout_gravity="top"
+                android:layout_marginTop="10dp"
+                android:layout_marginLeft="15dp"
+                android:layout_marginRight="15dp">
+            </ListView>
 
-        <com.google.android.material.card.MaterialCardView
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="3"
-
-            android:layout_marginRight="30dp"
-            android:layout_marginLeft="30dp"
-            android:layout_marginTop="20dp">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/confirmContact"
-                android:textSize="20sp"
-                android:layout_marginTop="20dp"
-                android:layout_marginStart="10dp">
-            </TextView>
         </com.google.android.material.card.MaterialCardView>
