+++ app/src/main/res/layout/fragment_create_study_step_five.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".createStudyFragment_StepFive"
-    android:background="?android:windowBackground">
+    android:background="?android:windowBackground"
+    tools:context=".createStudyFragment_StepFive">
 
 
     <LinearLayout
@@ -35,31 +35,41 @@
             android:id="@+id/pageDesc"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_weight="4"
-            android:textSize="20sp"
-            android:textAlignment="center"
             android:layout_marginLeft="5dp"
-            android:layout_marginRight="5dp"
             android:layout_marginTop="30dp"
-            android:text="@string/fragment_create_study_desc_page_five" />
-
-
-        <ListView
-            android:id="@+id/createDatelist"
-            android:layout_width="fill_parent"
-            android:layout_height="0dp"
-            android:layout_marginBottom="5dp"
+            android:layout_marginRight="5dp"
             android:layout_weight="3"
-            android:drawSelectorOnTop="false" />
+            android:text="@string/fragment_create_study_desc_page_five"
+            android:textAlignment="center"
+            android:textSize="20sp" />
 
-        <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-            android:id="@+id/fab"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:layout_gravity="bottom|right"
-            android:text="@string/fragment_create_study_pageFive_fab"
-            app:icon="@drawable/ic_baseline_add_24"/>
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:layout_weight="4"
+            android:orientation="vertical">
+
+            <ListView
+                android:id="@+id/createDatelist"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_marginBottom="5dp"
+                android:layout_marginStart="5dp"
+                android:drawSelectorOnTop="false" />
+
+            <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
+                android:id="@+id/fab"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_alignParentBottom="true"
+                android:layout_alignParentRight="true"
+                android:layout_marginBottom="16dp"
+                android:layout_marginRight="16dp"
+                android:text="@string/fragment_create_study_pageFive_fab"
+                app:icon="@drawable/ic_baseline_add_24" />
+
+        </RelativeLayout>
+
 
     </LinearLayout>
